refact/#46: CRUD URL 수정
- role 기준으로 API 및 컨트롤러 분리
